refactor(test): move more slow vector index tests to integration tests
Move test_create_index_nulls (10 parameterized cases) and
test_initialize_vector_index_empty_dataset to the integration
test module to further reduce unit test runtime.
